Added a new regex to allow decimal point numbers
Added the possibility to show decimals with parseFloat

Wrote first version of the rounding method for display

Added the format function for the decimals globally

Added some comments to remember the pending tasks on properly handling formatting with settings and spreading the decimal formatting through the app

+++ src/components/Dashboard/Dashboard.test.js
@@ -66,6 +66,6 @@ describe('Dashboard component test', () => {
     });
 
     const totalSum = DashboardInstance.getSum('expenses');
-    expect(totalSum).toBe(parseInt(newEntry.amount) * 4);
+    expect(totalSum).toBe(parseFloat(newEntry.amount) * 4);
   })
 })
